net: stmmac: Set DMA buffer size in HW
[ Upstream commit 4205c88eaf17b5f3ee30032d68df55cd5d9077a1 ]

This is clearly a bug.

We need to set the DMA buffer size in the HW otherwise corruption can
occur when receiving packets.

This is probably not occuring because of small MTU values and because HW
has a default value internally (which currently is bigger than default
buffer size).

Signed-off-by: Jose Abreu <joabreu@synopsys.com>
Cc: David S. Miller <davem@davemloft.net>
Cc: Joao Pinto <jpinto@synopsys.com>
Cc: Vitor Soares <soares@synopsys.com>
Cc: Giuseppe Cavallaro <peppe.cavallaro@st.com>
Cc: Alexandre Torgue <alexandre.torgue@st.com>
Signed-off-by: David S. Miller <davem@davemloft.net>
Change-Id: If6774834d137b72ca0460e37d2f17dfadb51b477
(cherry picked from commit 836f656deae7e524561bf517e710ebff82b57a51)