pci: dwc: semidrive: Remove pcie link up check
It will fail to enumerate the device by run reboot command in linux
shell. The root cause is that the pcie link status is not cleared.
but the PCIe device(mv5072) will be reset by a gpio in bootloader,
so the PCIe link status is wrong, we need reinitialize it, so
remove this check.

ID: 1782
Change-Id: I88db75d1b570d0f6f6ff68cdc6299ff5cc123456
Signed-off-by: dejin.zheng <dejin.zheng@semidrive.com>